+++ gradle.properties
@@ -13,7 +13,7 @@ fabric_version=0.106.1+1.21.3
 
 # Mod Properties
 mod_version=1.1.1
-maven_group=tk.estecka.cloth-gamerules
+maven_group=fr.estecka.cloth-gamerules
 archives_base_name=cloth-gamerules
 
 # Dependencies
