+++ src/main/resources/cloth-gamerules.mixins.json
@@ -1,6 +1,6 @@
 {
 	"required": true,
-	"package": "tk.estecka.clothgamerules.mixin",
+	"package": "fr.estecka.clothgamerules.mixin",
 	"compatibilityLevel": "JAVA_17",
 	"mixins": [
 		"CreateWorldMoreTabMixin"
